requirements.txt
requests==2.28.2
aiogram=2.24
python-dotenv==0.21.1
peewee==3.15.4
